fix warnings

+++ Source/AliveLib/PsxRender.hpp
@@ -13,7 +13,7 @@ EXPORT void CC PSX_ClearOTag_4F6290(int** otBuffer, int otBufferSize);
 
 EXPORT void CC PSX_TPage_Change_4F6430(__int16 tPage);
 EXPORT int CC PSX_EMU_SetDispType_4F9960(int dispType);
-EXPORT void CC PSX_EMU_Render_51EF90(__int16 x, __int16 y, int minX, int minY, BYTE r, BYTE g, BYTE b, __int16 w, __int16 h, int clut, int semiTrans);
+EXPORT void CC PSX_EMU_Render_51EF90(__int16 x, __int16 y, int minX, int minY, BYTE r, BYTE g, BYTE b, __int16 w, __int16 h, WORD clut, int semiTrans);
 
 ALIVE_VAR_EXTERN(int, sScreenXOffSet_BD30E4);
 ALIVE_VAR_EXTERN(int, sScreenYOffset_BD30A4);
